backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
motor
certifi
python-multipart
pyyaml
google-generativeai
anthropic